Refactor setPhysicsBody function with the extension
